fix(several files): Fix the home view, and create Row component to fix table in MemberTable component
