+++ src/HomeView/HomeView.jsx
@@ -1,5 +1,5 @@
-import React from "react";
-import { useState, useEffect } from "react";
+import React from 'react';
+import { useState, useEffect } from 'react';
 import {
   Box,
   Flex,
@@ -11,13 +11,12 @@ import {
   Stack,
   Text,
   useBreakpointValue,
-} from "@chakra-ui/react";
-import { useDataLayerValue } from "../DataLayer";
-import CardType1 from "../Shared/Cards/CardType1";
-import { colorGenerate } from "../Shared/Functions/changeBgColor";
-import CardType2 from "../Shared/Cards/CardType2";
-import { spotify } from "../App";
-import PlaylistTable from '../PlaylistTable/PlaylistTable';
+} from '@chakra-ui/react';
+import { useDataLayerValue } from '../DataLayer';
+import CardType1 from '../Shared/Cards/CardType1';
+import { colorGenerate } from '../Shared/Functions/changeBgColor';
+import CardType2 from '../Shared/Cards/CardType2';
+import { spotify } from '../App';
 
 const HomeView = () => {
   const [{ playlists, featuredPlaylists, recentlyPlaylists }, dispatch] =
@@ -25,7 +24,7 @@ const HomeView = () => {
   const [newMusic, setNewMusic] = useState({});
   const [idsList] = useState([]);
 
-  const [color, setColor] = useState("");
+  const [color, setColor] = useState('');
 
   const fivePlaylists = playlists?.items?.sort(
     (a, b) => b?.tracks?.total - a?.tracks?.total
@@ -33,7 +32,7 @@ const HomeView = () => {
 
   useEffect(() => {
     spotify
-      .getNewReleases({ market: "CO", limit: 10 })
+      .getNewReleases({ market: 'CO', limit: 10 })
       .then((data) => setNewMusic(data.albums));
 
     spotify
@@ -44,7 +43,7 @@ const HomeView = () => {
       .then((data) => {
         const list = data.items
           .filter((i) => i.context)
-          .map((l) => l.context.href.split("/")[5]);
+          .map((l) => l.context.href.split('/')[5]);
 
         return [...new Set(list)];
       })
@@ -53,7 +52,7 @@ const HomeView = () => {
           spotify.getPlaylist(l).then((recentlyPlaylists) => {
             idsList.push({ ...recentlyPlaylists });
             dispatch({
-              type: "SET_RECENTLY_PLAYLISTS",
+              type: 'SET_RECENTLY_PLAYLISTS',
               recentlyPlaylists: idsList,
             });
           })
@@ -70,7 +69,7 @@ const HomeView = () => {
     md: 3,
     lg: 4,
     xl: 5,
-    "2xl": 7,
+    '2xl': 7,
   });
 
   const currentHour = new Date();
@@ -96,21 +95,18 @@ const HomeView = () => {
       >
         <Heading color="white" fontSize="32px">
           {currentHour <= 12
-            ? "¡Buenos días!"
+            ? '¡Buenos días!'
             : currentHour.getHours() > 12 && currentHour.getHours() <= 18
-            ? "¡Buenas tardes!"
-            : "¡Buenas noches!"}
+            ? '¡Buenas tardes!'
+            : '¡Buenas noches!'}
         </Heading>
-        
-        </Box>
-        <PlaylistTable />
 
         <Grid
           w="100%"
           h="auto"
           mt={6}
-          templateColumns={{ base: "repeat(2, 1fr)", lg: "repeat(3, 1fr)" }}
-          templateRows={{ base: "repeat(3, 1fr)", lg: "repeat(2, 1fr)" }}
+          templateColumns={{ base: 'repeat(2, 1fr)', lg: 'repeat(3, 1fr)' }}
+          templateRows={{ base: 'repeat(3, 1fr)', lg: 'repeat(2, 1fr)' }}
           gap={4}
         >
           <GridItem>
@@ -228,4 +224,3 @@ const HomeView = () => {
 };
 
 export default HomeView;
-
